# config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Clerk Backend API",
    "x-logo": {
      "url": "https://clerk.com/_next/image?url=%2Fimages%2Fclerk-logo.svg&w=96&q=75",
      "altText": "Clerk docs",
      "href": "https://clerk.com/docs"
    },
    "contact": {
      "email": "support@clerk.com",
      "name": "Clerk Platform Team",
      "url": "https://clerk.com/support"
    },
    "description": "The Clerk REST Backend API, meant to be accessed by backend\nservers.\n\n### Versions\n\nWhen the API changes in a way that isn't compatible with older versions, a new version is released.\nEach version is identified by its release date, e.g. `2021-02-05`. For more information, please see [Clerk API Versions](https://clerk.com/docs/backend-requests/versioning/overview).\n\n\nPlease see https://clerk.com/docs for more information.",
    "version": "v1",
    "termsOfService": "https://clerk.com/terms",
    "license": {
      "name": "MIT",
      "url": "https://github.com/clerkinc/clerk-sdk-go/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://api.clerk.com/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/users/{user_id}": {
      "get": {
        "operationId": "GetUser",
        "summary": "Retrieve a user",
        "description": "Retrieve the details of a user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The ID of the user to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/User"
          },
          "400": {
            "$ref": "#/components/responses/ClerkErrors"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationInvalid"
          },
          "404": {
            "$ref": "#/components/responses/ResourceNotFound"
          }
        }
      },
      "patch": {
        "operationId": "UpdateUser",
        "summary": "Update a user",
        "description": "Update a user's attributes.\n\nYou can set the user's primary contact identifiers (email address and phone numbers) by updating the `primary_email_address_id` and `primary_phone_number_id` attributes respectively.\nBoth IDs should correspond to verified identifications that belong to the user.\n\nYou can remove a user's username by setting the username attribute to null or the blank string \"\".\nThis is a destructive action; the identification will be deleted forever.\nUsernames can be removed only if they are optional in your instance settings and there's at least one other identifier which can be used for authentication.\n\nThis endpoint allows changing a user's password. When passing the `password` parameter directly you have two further options.\nYou can ignore the password policy checks for your instance by setting the `skip_password_checks` parameter to `true`.\nYou can also choose to sign the user out of all their active sessions on any device once the password is updated. Just set `sign_out_of_other_sessions` to `true`.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The ID of the user to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "external_id": {
                    "type": "string",
                    "description": "The ID of the user as used in your external systems or your previous authentication solution.\nMust be unique across your instance.",
                    "nullable": true
                  },
                  "first_name": {
                    "type": "string",
                    "description": "The first name to assign to the user",
                    "nullable": true
                  },
                  "last_name": {
                    "type": "string",
                    "description": "The last name to assign to the user",
                    "nullable": true
                  },
                  "primary_email_address_id": {
                    "type": "string",
                    "description": "The ID of the email address to set as primary.\nIt must be verified, and present on the current user."
                  },
                  "notify_primary_email_address_changed": {
                    "type": "boolean",
                    "description": "If set to `true`, the user will be notified that their primary email address has changed.\nBy default, no notification is sent.",
                    "default": false
                  },
                  "primary_phone_number_id": {
                    "type": "string",
                    "description": "The ID of the phone number to set as primary.\nIt must be verified, and present on the current user."
                  },
                  "primary_web3_wallet_id": {
                    "type": "string",
                    "description": "The ID of the web3 wallets to set as primary.\nIt must be verified, and present on the current user."
                  },
                  "username": {
                    "type": "string",
                    "description": "The username to give to the user.\nIt must be unique across your instance.",
                    "nullable": true
                  },
                  "profile_image_id": {
                    "type": "string",
                    "description": "The ID of the image to set as the user's profile image",
                    "nullable": true
                  },
                  "password": {
                    "type": "string",
                    "description": "The plaintext password to give the user.\nMust be at least 8 characters long, and can not be in any list of hacked passwords.",
                    "nullable": true
                  },
                  "password_digest": {
                    "type": "string",
                    "description": "In case you already have the password digests and not the passwords, you can use them for the newly created user via this property.\nThe digests should be generated with one of the supported algorithms.\nThe hashing algorithm can be specified using the `password_hasher` property."
                  },
                  "password_hasher": {
                    "type": "string",
                    "description": "The hashing algorithm that was used to generate the password digest.\nThe algorithms we support at the moment are [bcrypt](https://en.wikipedia.org/wiki/Bcrypt), [bcrypt_sha256_django](https://docs.djangoproject.com/en/4.0/topics/auth/passwords/),\n[md5](https://en.wikipedia.org/wiki/MD5), pbkdf2_sha256, pbkdf2_sha512, [pbkdf2_sha256_django](https://docs.djangoproject.com/en/4.0/topics/auth/passwords/),\n[phpass](https://www.openwall.com/phpass/), [scrypt_firebase](https://firebaseopensource.com/projects/firebase/scrypt/),\n[sha256](https://en.wikipedia.org/wiki/SHA-2), [scrypt_werkzeug](https://werkzeug.palletsprojects.com/en/3.0.x/utils/#werkzeug.security.generate_password_hash)\nand the [argon2](https://argon2.online/) variants argon2i and argon2id.\n\nIf you need support for any particular hashing algorithm, [please let us know](https://clerk.com/support).\n\nNote: for password hashers considered insecure (at this moment MD5 and SHA256), the corresponding user password hashes will be transparently migrated to Bcrypt (a secure hasher) upon the user's first successful password sign in.\nInsecure schemes are marked with `(insecure)` in the list below.\n\nEach of the supported hashers expects the incoming digest to be in a particular format. Specifically:\n\n**bcrypt:** The digest should be of the following form:\n\n`$<algorithm version>$<cost>$<salt & hash>`\n\n**bcrypt_sha256_django:** This is the Django-specific variant of Bcrypt, using SHA256 hashing function. The format should be as follows (as exported from Django):\n\n`bcrypt_sha256$$<algorithm version>$<cost>$<salt & hash>`\n\n**md5** (insecure): The digest should follow the regular form e.g.:\n\n`5f4dcc3b5aa765d61d8327deb882cf99`\n\n**pbkdf2_sha256:** This is the PBKDF2 algorithm using the SHA256 hashing function. The format should be as follows:\n\n`pbkdf2_sha256$<iterations>$<salt>$<hash>`\n\nNote: Both the salt and the hash are expected to be base64-encoded.\n\n**pbkdf2_sha512:** This is the PBKDF2 algorithm using the SHA512 hashing function. The format should be as follows:\n\n`pbkdf2_sha512$<iterations>$<salt>$<hash>`\n\n  _iterations:_ The number of iterations used. Must be an integer less than 420000.\n  _salt:_ The salt used when generating the hash. Must be less than bytes.\n  _hash:_ The hex-encoded hash. Must have been generated with a key length less than 1024 bytes.\n\n**pbkdf2_sha256_django:** This is the Django-specific variant of PBKDF2 and the digest should have the following format (as exported from Django):\n\n`pbkdf2_sha256$<iterations>$<salt>$<hash>`\n\nNote: The salt is expected to be un-encoded, the hash is expected base64-encoded.\n\n**pbkdf2_sha1:** This is similar to pkbdf2_sha256_django, but with two differences:\n1. uses sha1 instead of sha256\n2. accepts the hash as a hex-encoded string\n\nThe format is the following:\n\n`pbkdf2_sha1$<iterations>$<salt>$<hash-as-hex-string>`\n\n**phpass:** Portable public domain password hashing framework for use in PHP applications. Digests hashed with phpass have the following sections:\n\nThe format is the following:\n\n`$P$<rounds><salt><encoded-checksum>`\n\n- $P$ is the prefix used to identify phpass hashes.\n- rounds is a single character encoding a 6-bit integer representing the number of rounds used.\n- salt is eight characters drawn from [./0-9A-Za-z], providing a 48-bit salt.\n- checksum is 22 characters drawn from the same set, encoding the 128-bit checksum with MD5.\n\n**scrypt_firebase:** The Firebase-specific variant of scrypt.\nThe value is expected to have 6 segments separated by the $ character and include the following information:\n\n_hash:_ The actual Base64 hash. This can be retrieved when exporting the user from Firebase.\n_salt:_ The salt used to generate the above hash. Again, this is given when exporting the user.\n_signer key:_ The base64 encoded signer key.\n_salt separator:_ The base64 encoded salt separator.\n_rounds:_ The number of rounds the algorithm needs to run.\n_memory cost:_ The cost of the algorithm run\n\nThe first 2 (hash and salt) are per user and can be retrieved when exporting the user from Firebase.\nThe other 4 values (signer key, salt separator, rounds and memory cost) are project-wide settings and can be retrieved from the project's password hash parameters.\n\nOnce you have all these, you can combine it in the following format and send this as the digest in order for Clerk to accept it:\n\n`<hash>$<salt>$<signer key>$<salt separator>$<rounds>$<memory cost>`\n\n**scrypt_werkzeug:** The Werkzeug-specific variant of scrypt.\n\nThe value is expected to have 3 segments separated by the $ character and include the following information:\n\n_algorithm args:_ The algorithm used to generate the hash.\n_salt:_ The salt used to generate the above hash.\n_hash:_ The actual Base64 hash.\n\nThe algorithm args are the parameters used to generate the hash and are included in the digest.\n\n**argon2i:** Algorithms in the argon2 family generate digests that encode the following information:\n\n_version (v):_ The argon version, version 19 is assumed\n_memory (m):_ The memory used by the algorithm (in kibibytes)\n_iterations (t):_ The number of iterations to perform\n_parallelism (p):_ The number of threads to use\n\nParts are demarcated by the `$` character, with the first part identifying the algorithm variant.\nThe middle part is a comma-separated list of the encoding options (memory, iterations, parallelism).\nThe final part is the actual digest.\n\n`$argon2i$v=19$m=4096,t=3,p=1$4t6CL3P7YiHBtwESXawI8Hm20zJj4cs7/4/G3c187e0$m7RQFczcKr5bIR0IIxbpO2P0tyrLjf3eUW3M3QSwnLc`\n\n**argon2id:** See the previous algorithm for an explanation of the formatting.\n\nFor the argon2id case, the value of the algorithm in the first part of the digest is `argon2id`:\n\n`$argon2id$v=19$m=64,t=4,p=8$Z2liZXJyaXNo$iGXEpMBTDYQ8G/71tF0qGjxRHEmR3gpGULcE93zUJVU`\n\n**sha256** (insecure): The digest should be a 64-length hex string, e.g.:\n\n`9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08`",
                    "enum": [
                      "argon2i",
                      "argon2id",
                      "bcrypt",
                      "bcrypt_sha256_django",
                      "md5",
                      "pbkdf2_sha256",
                      "pbkdf2_sha512",
                      "pbkdf2_sha256_django",
                      "pbkdf2_sha1",
                      "phpass",
                      "scrypt_firebase",
                      "scrypt_werkzeug",
                      "sha256"
                    ]
                  },
                  "skip_password_checks": {
                    "type": "boolean",
                    "description": "Set it to `true` if you're updating the user's password and want to skip any password policy settings check. This parameter can only be used when providing a `password`.",
                    "nullable": true
                  },
                  "sign_out_of_other_sessions": {
                    "type": "boolean",
                    "description": "Set to `true` to sign out the user from all their active sessions once their password is updated. This parameter can only be used when providing a `password`.",
                    "nullable": true
                  },
                  "totp_secret": {
                    "type": "string",
                    "description": "In case TOTP is configured on the instance, you can provide the secret to enable it on the specific user without the need to reset it.\nPlease note that currently the supported options are:\n* Period: 30 seconds\n* Code length: 6 digits\n* Algorithm: SHA1"
                  },
                  "backup_codes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "If Backup Codes are configured on the instance, you can provide them to enable it on the specific user without the need to reset them.\nYou must provide the backup codes in plain format or the corresponding bcrypt digest."
                  },
                  "public_metadata": {
                    "type": "object",
                    "description": "Metadata saved on the user, that is visible to both your Frontend and Backend APIs"
                  },
                  "private_metadata": {
                    "type": "object",
                    "description": "Metadata saved on the user, that is only visible to your Backend API"
                  },
                  "unsafe_metadata": {
                    "type": "object",
                    "description": "Metadata saved on the user, that can be updated from both the Frontend and Backend APIs.\nNote: Since this data can be modified from the frontend, it is not guaranteed to be safe."
                  },
                  "delete_self_enabled": {
                    "type": "boolean",
                    "description": "If true, the user can delete themselves with the Frontend API.",
                    "nullable": true
                  },
                  "create_organization_enabled": {
                    "type": "boolean",
                    "description": "If true, the user can create organizations with the Frontend API.",
                    "nullable": true
                  },
                  "created_at": {
                    "type": "string",
                    "description": "A custom date/time denoting _when_ the user signed up to the application, specified in RFC3339 format (e.g. `2012-10-20T07:15:20.902Z`)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/User"
          },
          "400": {
            "$ref": "#/components/responses/ClerkErrors"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationInvalid"
          },
          "404": {
            "$ref": "#/components/responses/ResourceNotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      },
      "delete": {
        "operationId": "DeleteUser",
        "summary": "Delete a user",
        "description": "Delete the specified user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The ID of the user to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/DeletedObject"
          },
          "400": {
            "$ref": "#/components/responses/ClerkErrors"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationInvalid"
          },
          "404": {
            "$ref": "#/components/responses/ResourceNotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "schemas": {
...
      "ClerkError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "long_message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "meta": {
            "type": "object"
          },
          "clerk_trace_id": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "long_message",
          "code"
        ]
      },
      "ClerkErrors": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClerkError"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "errors"
        ]
      },
...
      "EmailAddress": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "description": "String representing the object's type. Objects of the same type share the same value.\n",
            "enum": [
              "email_address"
            ]
          },
          "email_address": {
            "type": "string"
          },
          "reserved": {
            "type": "boolean"
          },
          "verification": {
            "type": "object",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/OTP"
              },
              {
                "$ref": "#/components/schemas/Admin"
              },
              {
                "$ref": "#/components/schemas/Oauth"
              }
            ]
          },
          "linked_to": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdentificationLink"
            }
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation\n"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation\n"
          }
        },
        "required": [
          "object",
          "email_address",
          "verification",
          "linked_to",
          "reserved",
          "created_at",
          "updated_at"
        ]
      },
      "DeletedObject": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "object": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "object",
          "deleted"
        ]
      },
      "PhoneNumber": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "description": "String representing the object's type. Objects of the same type share the same value.\n",
            "enum": [
              "phone_number"
            ]
          },
          "phone_number": {
            "type": "string"
          },
          "reserved_for_second_factor": {
            "type": "boolean"
          },
          "default_second_factor": {
            "type": "boolean"
          },
          "reserved": {
            "type": "boolean"
          },
          "verification": {
            "type": "object",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/OTP"
              },
              {
                "$ref": "#/components/schemas/Admin"
              }
            ]
          },
          "linked_to": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdentificationLink"
            }
          },
          "backup_codes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation\n"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation\n"
          }
        },
        "required": [
          "object",
          "phone_number",
          "verification",
          "linked_to",
          "reserved",
          "created_at",
          "updated_at"
        ]
      },
...
      "Web3Wallet": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "description": "String representing the object's type. Objects of the same type share the same value.\n",
            "enum": [
              "web3_wallet"
            ]
          },
          "web3_wallet": {
            "type": "string"
          },
          "verification": {
            "type": "object",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Web3Signature"
              },
              {
                "$ref": "#/components/schemas/Admin"
              }
            ]
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation\n"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation\n"
          }
        },
        "required": [
          "object",
          "web3_wallet",
          "verification",
          "created_at",
          "updated_at"
        ]
      },
...
      "schemas-Passkey": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "description": "String representing the object's type. Objects of the same type share the same value.\n",
            "enum": [
              "passkey"
            ]
          },
          "name": {
            "type": "string"
          },
          "last_used_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of when the passkey was last used.\n"
          },
          "verification": {
            "type": "object",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Passkey"
              }
            ]
          }
        },
        "required": [
          "object",
          "name",
          "last_used_at",
          "verification"
        ]
      },
...
      "SAMLAccount": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "description": "String representing the object's type. Objects of the same type share the same value.\n",
            "enum": [
              "saml_account"
            ]
          },
          "provider": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "email_address": {
            "type": "string"
          },
          "first_name": {
            "type": "string",
            "nullable": true
          },
          "last_name": {
            "type": "string",
            "nullable": true
          },
          "provider_user_id": {
            "type": "string",
            "nullable": true
          },
          "public_metadata": {
            "type": "object"
          },
          "verification": {
            "type": "object",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/SAML"
              },
              {
                "$ref": "#/components/schemas/Ticket"
              }
            ]
          }
        },
        "required": [
          "id",
          "object",
          "provider",
          "active",
          "email_address",
          "verification"
        ]
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "description": "String representing the object's type. Objects of the same type share the same value.\n",
            "enum": [
              "user"
            ]
          },
          "external_id": {
            "nullable": true,
            "type": "string"
          },
          "primary_email_address_id": {
            "nullable": true,
            "type": "string"
          },
          "primary_phone_number_id": {
            "nullable": true,
            "type": "string"
          },
          "primary_web3_wallet_id": {
            "nullable": true,
            "type": "string"
          },
          "username": {
            "nullable": true,
            "type": "string"
          },
          "first_name": {
            "nullable": true,
            "type": "string"
          },
          "last_name": {
            "nullable": true,
            "type": "string"
          },
          "profile_image_url": {
            "type": "string",
            "deprecated": true
          },
          "image_url": {
            "type": "string"
          },
          "has_image": {
            "type": "boolean"
          },
          "public_metadata": {
            "type": "object"
          },
          "private_metadata": {
            "nullable": true,
            "type": "object"
          },
          "unsafe_metadata": {
            "type": "object"
          },
          "email_addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmailAddress"
            }
          },
          "phone_numbers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhoneNumber"
            }
          },
          "web3_wallets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Web3Wallet"
            }
          },
          "passkeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/schemas-Passkey"
            }
          },
          "password_enabled": {
            "type": "boolean"
          },
          "two_factor_enabled": {
            "type": "boolean"
          },
          "totp_enabled": {
            "type": "boolean"
          },
          "backup_code_enabled": {
            "type": "boolean"
          },
          "mfa_enabled_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of when MFA was last enabled for this user. It should be noted that this field is not nullified if MFA is disabled.\n",
            "nullable": true
          },
          "mfa_disabled_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of when MFA was last disabled for this user. It should be noted that this field is not nullified if MFA is enabled again.\n",
            "nullable": true
          },
          "external_accounts": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "saml_accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SAMLAccount"
            }
          },
          "last_sign_in_at": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Unix timestamp of last sign-in.\n"
          },
          "banned": {
            "type": "boolean",
            "description": "Flag to denote whether user is banned or not.\n"
          },
          "locked": {
            "type": "boolean",
            "description": "Flag to denote whether user is currently locked, i.e. restricted from signing in or not.\n"
          },
          "lockout_expires_in_seconds": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "The number of seconds remaining until the lockout period expires for a locked user. A null value for a locked user indicates that lockout never expires.\n"
          },
          "verification_attempts_remaining": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "The number of verification attempts remaining until the user is locked. Null if account lockout is not enabled. Note: if a user is locked explicitly via the Backend API, they may still have verification attempts remaining.\n"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of last update.\n"
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation.\n"
          },
          "delete_self_enabled": {
            "type": "boolean",
            "description": "If enabled, user can delete themselves via FAPI.\n"
          },
          "create_organization_enabled": {
            "type": "boolean",
            "description": "If enabled, user can create organizations via FAPI.\n"
          },
          "last_active_at": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Unix timestamp of the latest session activity, with day precision.\n",
            "example": 1700690400000
          }
        }
      },
...
    },
    "responses": {
...
      "ClerkErrors": {
        "description": "Request was not successful",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ClerkErrors"
            }
          }
        }
      },
      "AuthenticationInvalid": {
        "description": "Authentication invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ClerkErrors"
            }
          }
        }
      },
...
      "UnprocessableEntity": {
        "description": "Invalid request parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ClerkErrors"
            }
          }
        }
      },
...
      "ResourceNotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ClerkErrors"
            }
          }
        }
      },
...
      "DeletedObject": {
        "description": "Deleted Object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DeletedObject"
            }
          }
        }
      },
...
      "User": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
...
    }
  }
}